requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
# Ingestion: beliebige Formate → PNG für Gemini
pymupdf>=1.24.0
Pillow>=10.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
# Gemini 3 Flash (Dokumentenextraktion)
google-genai>=1.0.0
# API (Render)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
supabase>=2.0.0
httpx>=0.26.0
